메인 비즈니스 로직 상세페이지 완성

+++ src/main/java/project7/tulipmetric/domain/Market/Market.java
@@ -27,7 +27,7 @@ public class Market {
     private String chart; //12 개월 코스피 지수 변동.
 
     @Column(nullable = false)
-    private int growthRate30d;
+    private double growthRate30d;
 
     @Column(length = 20,nullable = false)
     private String marketStatus;    // crashed/declining/stable/growing/overvalued
